news_feeds/gdelt_news/requirements.txt
requests==2.32.3
PyYAML==6.0.2
tenacity==8.5.0
